Harshaldev

+++ resources/views/frontend_admin/role.blade.php
@@ -47,27 +47,31 @@
                                         </tr>
                                     </thead>
                                     <tbody>
-                                        @foreach($roles as $key => $role)
-                                        <tr>
-                                            <td>{{ $key + 1 }}</td>
-                                            <td>{{ $role->role_name }}</td>
-                                            <td>
-                                                <!-- Edit action link with encrypted ID -->
-                                                <button class="btn btn-warning toggle-edit-form"
-                                                    data-role-id="{{ $role->tbl_role_id }}"
-                                                    data-encrypted-id="{{ $role->encrypted_id }}">Edit</button>
+                                        @php
+                                            $serialNumber = 0;
+                                        @endphp
+                                        @foreach($roles as $role)
+                                            @if($role->tbl_role_id !== 1)
+                                                <tr>
+                                                    <td>{{ ++$serialNumber }}</td>
+                                                    <td>{{ $role->role_name }}</td>
+                                                    <td>
+                                                        <!-- Edit action link with encrypted ID -->
+                                                        <button class="btn btn-warning toggle-edit-form"
+                                                                data-role-id="{{ $role->tbl_role_id }}"
+                                                                data-encrypted-id="{{ $role->encrypted_id }}">Edit</button>
 
-                                                
-                                                <!-- Delete action form with encrypted ID -->
-                                                <a href="/admin/deleterole/{{$role->encrypted_id}}" class="btn btn-danger delete-role" data-encrypted-id="{{ $role->encrypted_id }}">Delete</a>
-                                                
-                                                <!-- Assign Module action link -->
-                                                <a href="/admin/assignmodule" class="btn btn-info">Assign Module</a>
+                                                        <!-- Delete action form with encrypted ID -->
+                                                        <a href="/admin/deleterole/{{$role->encrypted_id}}" class="btn btn-danger delete-role" data-encrypted-id="{{ $role->encrypted_id }}">Delete</a>
 
-                                            </td>
-                                        </tr>
+                                                        <!-- Assign Module action link -->
+                                                        <a href="/admin/assignmodule/{{$role->encrypted_id}}" class="btn btn-info">Assign Module</a>
+                                                    </td>
+                                                </tr>
+                                            @endif
                                         @endforeach
                                     </tbody>
+
                                 </table>
                             </div>
                         </div>
@@ -83,7 +87,7 @@
 <div class="modal fade" id="addRoleModal" tabindex="-1" role="dialog" aria-labelledby="addRoleModalLabel" aria-hidden="true">
     <div class="modal-dialog" role="document">
         <div class="modal-content">
-            <form action="/admin/addrole" method="POST">
+            <form action="/admin/storerole" method="POST">
                 @csrf
                 <div class="modal-header">
                     <h5 class="modal-title" id="addRoleModalLabel">Add New Role</h5>
